+++ plugin-gradle/src/test/java/com/diffplug/gradle/spotless/JavaEclipseTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2024 DiffPlug
+ * Copyright 2016-2025 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import org.junit.jupiter.api.Test;
 
 class JavaEclipseTest extends GradleIntegrationHarness {
 	@Test
-	void settingsWithContentWithoutFile() throws IOException {
+	void settingsWithProprtiesContent() throws IOException {
 		setFile("build.gradle").toLines(
 				"plugins {",
 				"  id 'com.diffplug.spotless'",
@@ -32,6 +32,29 @@ class JavaEclipseTest extends GradleIntegrationHarness {
 				"spotless {",
 				"  java {  eclipse().configProperties(\"\"\"",
 				"valid_line_oriented.prefs.string=string",
+				"\"\"\")  }",
+				"}");
+
+		gradleRunner().withArguments("spotlessApply").build();
+	}
+
+	@Test
+	void settingsWithXmlContent() throws IOException {
+		setFile("build.gradle").toLines(
+				"plugins {",
+				"  id 'com.diffplug.spotless'",
+				"  id 'java'",
+				"}",
+				"repositories { mavenCentral() }",
+				"",
+				"spotless {",
+				"  java {  eclipse().configXml(\"\"\"",
+				"<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"no\"?>",
+				"<profiles version=\"12\">",
+				"  <profile kind=\"CodeFormatterProfile\" name=\"Spotless\" version=\"12\">",
+				"    <setting id=\"valid_line_oriented.prefs.string\" value=\"string\" />",
+				"  </profile>",
+				"</profiles>",
 				"\"\"\")  }",
 				"}");